rename service files; update user info implementation
